Refactor to maintain model and ui project structure
