+++ app/src/main/java/com/example/planlekcji/MainActivity.java
@@ -10,7 +10,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.app.AppCompatDelegate;
 import androidx.viewpager2.widget.ViewPager2;
 
-import com.example.planlekcji.fragments.ViewPagerAdapter;
+import com.example.planlekcji.fragments.model.ViewPagerAdapter;
 import com.example.planlekcji.utils.ToastUtils;
 
 import com.google.android.material.tabs.TabLayout;
