+++ app/src/main/java/com/example/planlekcji/fragments/model/ViewPagerAdapter.java
@@ -1,9 +1,14 @@
-package com.example.planlekcji.fragments;
+package com.example.planlekcji.fragments.model;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentActivity;
 import androidx.viewpager2.adapter.FragmentStateAdapter;
+
+import com.example.planlekcji.fragments.ui.HomeFragment;
+import com.example.planlekcji.fragments.ui.ReplacementsFragment;
+import com.example.planlekcji.fragments.ui.SettingsFragment;
+import com.example.planlekcji.fragments.ui.TimetableFragment;
 
 public class ViewPagerAdapter extends FragmentStateAdapter {
     public ViewPagerAdapter(@NonNull FragmentActivity fragmentActivity) {
