+++ app/src/main/java/com/example/planlekcji/fragments/ui/TimetableFragment.java
@@ -1,4 +1,4 @@
-package com.example.planlekcji.fragments;
+package com.example.planlekcji.fragments.ui;
 
 import android.os.Bundle;
