+++ app/src/main/res/layout/fragment_home.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.HomeFragment">
+    tools:context=".fragments.ui.HomeFragment">
 
     <TextView
         android:layout_width="match_parent"
