+++ app/src/main/res/layout/fragment_replacements.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.ReplacementsFragment">
+    tools:context=".fragments.ui.ReplacementsFragment">
 
     <ScrollView
         android:id="@+id/scrollView_replacements"
